+++ plugins/source/aws/resources/services/cloudwatchlogs/log_groups.go
@@ -1,6 +1,9 @@
 package cloudwatchlogs
 
 import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/service/cloudwatchlogs"
 	"github.com/aws/aws-sdk-go-v2/service/cloudwatchlogs/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
@@ -31,5 +34,33 @@ func LogGroups() *schema.Table {
 				Resolver: resolveLogGroupTags,
 			},
 		},
+		Relations: []*schema.Table{
+			subscriptionFilters(),
+		},
 	}
+}
+
+func fetchCloudwatchlogsLogGroups(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	var config cloudwatchlogs.DescribeLogGroupsInput
+	c := meta.(*client.Client)
+	svc := c.Services().Cloudwatchlogs
+	paginator := cloudwatchlogs.NewDescribeLogGroupsPaginator(svc, &config)
+	for paginator.HasMorePages() {
+		page, err := paginator.NextPage(ctx)
+		if err != nil {
+			return err
+		}
+		res <- page.LogGroups
+	}
+	return nil
+}
+
+func resolveLogGroupTags(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
+	lg := resource.Item.(types.LogGroup)
+	svc := meta.(*client.Client).Services().Cloudwatchlogs
+	out, err := svc.ListTagsLogGroup(ctx, &cloudwatchlogs.ListTagsLogGroupInput{LogGroupName: lg.LogGroupName})
+	if err != nil {
+		return err
+	}
+	return resource.Set(c.Name, out.Tags)
 }
